Squashed commit of the following:
commit ffba52b898c49caeccc3d4d57ff3e675203df355
Author: Ana Gavrilovic <gavrilovic.ra65.2018@uns.ac.rs>
Date:   Thu Jan 13 18:21:44 2022 +0100

    refactor: user version

commit 6c25d3ac0caf7cbfee8865536cb4c0cd78f613f2
Author: Ana Gavrilovic <gavrilovic.ra65.2018@uns.ac.rs>
Date:   Thu Jan 13 18:14:26 2022 +0100

    feat: konflikta situacija 2

commit ff5882d857c11d2c50c7d1b7b6dff5382c8fb1b2
Author: Ana Gavrilovic <gavrilovic.ra65.2018@uns.ac.rs>
Date:   Thu Jan 13 16:25:01 2022 +0100

    feat: transakcije

commit 881ea0a33c37cbd36e27435c726f7ed3b3ecdf02
Author: Ana Gavrilovic <gavrilovic.ra65.2018@uns.ac.rs>
Date:   Thu Jan 13 05:03:56 2022 +0100

    feat: konkurentan pristup bazi

+++ FishingBooker/backend/src/main/java/com/backend/service/DeleteRequestService.java
@@ -6,6 +6,7 @@ import net.bytebuddy.asm.Advice;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
@@ -25,19 +26,21 @@ public class DeleteRequestService {
     @Autowired
     private DeleteUserService deleteUserService;
 
-
+    @Transactional(readOnly = false)
     public void save(String email,String content){
         RegisteredUser user=userRepository.findByEmail(email);
         DeleteRequest deleteRequest = new DeleteRequest(user,content);
         deleteRequestRepository.save(deleteRequest);
     }
 
+    @Transactional(readOnly = true)
     public List<DeleteRequest> getAllRequests() {
         return deleteRequestRepository.findAll();
     }
 
+    @Transactional(readOnly = false)
     public void rejectDeleteRequest(Integer requestId, String response) {
-        DeleteRequest deleteRequest = deleteRequestRepository.findById(requestId).get();
+        DeleteRequest deleteRequest = deleteRequestRepository.findOneById(requestId);
         if (deleteRequest == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "No such request");
 
         deleteRequestRepository.delete(deleteRequest);
@@ -52,10 +55,12 @@ public class DeleteRequestService {
         emailService.sendSimpleMessage(email, "Delete Request Rejected", stringBuilder.toString());
     }
 
+    @Transactional(readOnly = false)
     public void approveDeleteRequest(Integer id) {
-        DeleteRequest deleteRequest = deleteRequestRepository.findById(id).get();
-        RegisteredUser registeredUser = deleteRequest.getRegisteredUser();
+        DeleteRequest deleteRequest = deleteRequestRepository.findOneById(id);
+        if (deleteRequest == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "No such request");
 
+        RegisteredUser registeredUser = deleteRequest.getRegisteredUser();
         deleteUserService.deleteUser(registeredUser.getId());
         sendApprovalEmail(registeredUser.getEmail());
     }
